fix: refactored generic models
